Port image command to new cli structure
* Port image command
* Add shared watch package
