+++ internal/cli/internal/command/deploy/deploy.go
@@ -4,35 +4,24 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
 	"path/filepath"
-	"reflect"
 	"strings"
-	"sync"
-	"time"
 
-	"github.com/briandowns/spinner"
 	"github.com/dustin/go-humanize"
 	"github.com/spf13/cobra"
-	"golang.org/x/sync/errgroup"
 
-	"github.com/superfly/flyctl/cmd/presenters"
 	"github.com/superfly/flyctl/pkg/iostreams"
-	"github.com/superfly/flyctl/pkg/logs"
 
 	"github.com/superfly/flyctl/api"
 	"github.com/superfly/flyctl/internal/build/imgsrc"
 	"github.com/superfly/flyctl/internal/cli/internal/app"
 	"github.com/superfly/flyctl/internal/cli/internal/command"
-	"github.com/superfly/flyctl/internal/cli/internal/config"
 	"github.com/superfly/flyctl/internal/cli/internal/flag"
-	"github.com/superfly/flyctl/internal/cli/internal/format"
 	"github.com/superfly/flyctl/internal/cli/internal/render"
 	"github.com/superfly/flyctl/internal/cli/internal/state"
+	"github.com/superfly/flyctl/internal/cli/internal/watch"
 	"github.com/superfly/flyctl/internal/client"
 	"github.com/superfly/flyctl/internal/cmdutil"
-	"github.com/superfly/flyctl/internal/deployment"
-	"github.com/superfly/flyctl/internal/flyerr"
 	"github.com/superfly/flyctl/internal/logger"
 )
 
@@ -45,7 +34,7 @@ func New() (cmd *cobra.Command) {
 
 	cmd = command.New("deploy [WORKING_DIRECTORY]", short, long, run,
 		command.RequireSession,
-		swapWorkingDirectoryIfRequired,
+		command.SwapWorkingDirectoryIfRequired,
 		command.RequireAppName,
 	)
 
@@ -95,27 +84,6 @@ func New() (cmd *cobra.Command) {
 	return
 }
 
-func swapWorkingDirectoryIfRequired(ctx context.Context) (context.Context, error) {
-	wd := flag.FirstArg(ctx)
-	if wd == "" {
-		return ctx, nil
-	}
-
-	if !filepath.IsAbs(wd) {
-		p, err := filepath.Abs(wd)
-		if err != nil {
-			return nil, fmt.Errorf("failed converting %s to an absolute path: %w", wd, err)
-		}
-		wd = p
-	}
-
-	if err := os.Chdir(wd); err != nil {
-		return nil, fmt.Errorf("failed changing working directory: %w", err)
-	}
-
-	return state.WithWorkingDirectory(ctx, wd), nil
-}
-
 func run(ctx context.Context) error {
 	appConfig, err := determineAppConfig(ctx)
 	if err != nil {
@@ -151,7 +119,7 @@ func run(ctx context.Context) error {
 		tb := render.NewTextBlock(ctx, fmt.Sprintf("Release command detected: %s\n", releaseCommand.Command))
 		tb.Done("This release will not be available until the release command succeeds.")
 
-		if err := watchReleaseCommand(ctx, releaseCommand.ID); err != nil {
+		if err := watch.WatchReleaseCommand(ctx, releaseCommand.ID); err != nil {
 			return err
 		}
 	}
@@ -163,7 +131,7 @@ func run(ctx context.Context) error {
 		return nil
 	}
 
-	err = watchDeployment(ctx)
+	err = watch.WatchDeployment(ctx)
 
 	return err
 }
@@ -357,285 +325,4 @@ func createRelease(ctx context.Context, appConfig *app.Config, img *imgsrc.Deplo
 	}
 
 	return release, releaseCommand, err
-}
-
-func watchReleaseCommand(ctx context.Context, id string) error {
-	g, ctx := errgroup.WithContext(ctx)
-	io := iostreams.FromContext(ctx)
-	client := client.FromContext(ctx).API()
-	interactive := io.IsInteractive()
-	appName := app.NameFromContext(ctx)
-
-	s := spinner.New(spinner.CharSets[11], 100*time.Millisecond)
-	s.Writer = os.Stderr
-	s.Prefix = "Running release task..."
-
-	if interactive {
-		s.Start()
-		defer s.Stop()
-	}
-
-	rcUpdates := make(chan api.ReleaseCommand)
-
-	var once sync.Once
-
-	startLogs := func(ctx context.Context, vmid string) {
-		once.Do(func() {
-			g.Go(func() error {
-				childCtx, cancel := context.WithCancel(ctx)
-				defer cancel()
-
-				opts := &logs.LogOptions{
-					MaxBackoff: time.Second,
-					AppName:    appName,
-					VMID:       vmid,
-				}
-
-				ls, err := logs.NewPollingStream(childCtx, client, opts)
-				if err != nil {
-					return err
-				}
-
-				for entry := range ls.Stream(childCtx, opts) {
-					if interactive {
-						s.Stop()
-						defer s.Start()
-					}
-
-					fmt.Println("\t", entry.Message)
-
-					// watch for the shutdown message
-					if entry.Message == "Starting clean up." {
-						cancel()
-					}
-				}
-
-				if err = ls.Err(); errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
-					err = nil
-				}
-				return err
-			})
-		})
-	}
-
-	g.Go(func() error {
-		var lastValue *api.ReleaseCommand
-		var errorCount int
-		defer close(rcUpdates)
-
-		for {
-			rc, err := client.GetReleaseCommand(ctx, id)
-			if err != nil {
-				errorCount += 1
-				if errorCount < 3 {
-					continue
-				}
-				return err
-			}
-
-			if !reflect.DeepEqual(lastValue, rc) {
-				lastValue = rc
-				rcUpdates <- *rc
-			}
-
-			if !rc.InProgress {
-				break
-			}
-
-			time.Sleep(500 * time.Millisecond)
-		}
-
-		return nil
-	})
-
-	g.Go(func() error {
-		// The logs goroutine will stop itself when it sees a shutdown log message.
-		// If the message never comes (delayed logs, etc) the deploy will hang.
-		// This timeout makes sure they always stop a few seconds after the release task is done.
-		logsCtx, logsCancel := context.WithCancel(ctx)
-		defer time.AfterFunc(3*time.Second, logsCancel)
-
-		for rc := range rcUpdates {
-			if interactive {
-				s.Prefix = fmt.Sprintf("Running release task (%s)...", rc.Status)
-			}
-
-			if rc.InstanceID != nil {
-				startLogs(logsCtx, *rc.InstanceID)
-			}
-
-			if !rc.InProgress && rc.Failed {
-				if rc.Succeeded && interactive {
-					s.FinalMSG = "Running release task...Done\n"
-				} else if rc.Failed {
-					return errors.New("Release command failed, deployment aborted")
-				}
-			}
-		}
-
-		return nil
-	})
-
-	return g.Wait()
-}
-
-func watchDeployment(ctx context.Context) error {
-	tb := render.NewTextBlock(ctx, "Monitoring deployment")
-
-	io := iostreams.FromContext(ctx)
-	appName := app.NameFromContext(ctx)
-	client := client.FromContext(ctx).API()
-	endmessage := ""
-
-	monitor := deployment.NewDeploymentMonitor(client, appName)
-
-	monitor.DeploymentStarted = func(idx int, d *api.DeploymentStatus) error {
-		if idx > 0 {
-			tb.Println()
-		}
-		tb.Println(format.DeploymentSummary(d))
-
-		return nil
-	}
-
-	// TODO check we aren't asking for JSON
-	monitor.DeploymentUpdated = func(d *api.DeploymentStatus, updatedAllocs []*api.AllocationStatus) error {
-		if io.IsInteractive() {
-			tb.Overwrite()
-
-			tb.Println(format.DeploymentAllocSummary(d))
-		} else {
-			for _, alloc := range updatedAllocs {
-				tb.Println(format.AllocSummary(alloc))
-			}
-		}
-
-		return nil
-	}
-
-	monitor.DeploymentFailed = func(d *api.DeploymentStatus, failedAllocs []*api.AllocationStatus) error {
-		// cmdCtx.Statusf("deploy", cmdctx.SDETAIL, "v%d %s - %s\n", d.Version, d.Status, d.Description)
-
-		if endmessage == "" && d.Status == "failed" {
-			if strings.Contains(d.Description, "no stable release to revert to") {
-				endmessage = fmt.Sprintf("v%d %s - %s\n", d.Version, d.Status, d.Description)
-			} else {
-				endmessage = fmt.Sprintf("v%d %s - %s and deploying as v%d \n", d.Version, d.Status, d.Description, d.Version+1)
-			}
-		}
-
-		if len(failedAllocs) > 0 {
-			tb.Println("Failed Instances")
-
-			x := make(chan *api.AllocationStatus)
-			var wg sync.WaitGroup
-			wg.Add(len(failedAllocs))
-
-			for _, a := range failedAllocs {
-				a := a
-				go func() {
-					defer wg.Done()
-					alloc, err := client.GetAllocationStatus(ctx, appName, a.ID, 30)
-					if err != nil {
-						//cmdCtx.Status("deploy", cmdctx.SERROR, "Error fetching alloc", a.ID, err)
-						tb.Printf("failed fetching alloc %s: %s", a.ID, err)
-
-						return
-					}
-					x <- alloc
-				}()
-			}
-
-			go func() {
-				wg.Wait()
-				close(x)
-			}()
-
-			var count int
-			for alloc := range x {
-				count++
-
-				tb.Println()
-				tb.Printf("Failure #%d\n", count)
-				tb.Println()
-
-				as, err := renderAllocationStatus(nil, alloc)
-				if err != nil {
-					return fmt.Errorf("failed rendering alloc status: %w", err)
-				}
-				tb.Println(as)
-
-				re, err := renderRecentEvents(alloc.Events)
-				if err != nil {
-					return fmt.Errorf("failed rendering recent events: %w", err)
-				}
-				tb.Println(re)
-
-				renderLogs(ctx, alloc)
-			}
-		}
-
-		return nil
-	}
-
-	monitor.DeploymentSucceeded = func(d *api.DeploymentStatus) error {
-		tb.Donef("v%d deployed successfully\n", d.Version)
-
-		return nil
-	}
-
-	monitor.Start(ctx)
-
-	if err := monitor.Error(); err != nil {
-		return err
-	}
-
-	if endmessage != "" {
-		tb.Done(endmessage)
-	}
-
-	if !monitor.Success() {
-		tb.Done("Troubleshooting guide at https://fly.io/docs/getting-started/troubleshooting/")
-		return flyerr.ErrAbort
-	}
-
-	return nil
-}
-
-func renderAllocationStatus(br []api.Region, s ...*api.AllocationStatus) (string, error) {
-	var sb strings.Builder
-	err := render.AllocationStatuses(&sb, "Instance", br, s...)
-
-	return sb.String(), err
-}
-
-func renderRecentEvents(events []api.AllocationEvent) (string, error) {
-	var sb strings.Builder
-	err := render.AllocationEvents(&sb, "Recent Events", events...)
-
-	return sb.String(), err
-}
-
-func renderLogs(ctx context.Context, alloc *api.AllocationStatus) {
-	out := iostreams.FromContext(ctx).Out
-	cfg := config.FromContext(ctx).JSONOutput
-
-	logPresenter := presenters.LogPresenter{
-		HideAllocID:    true,
-		HideRegion:     true,
-		RemoveNewlines: true,
-	}
-
-	for _, e := range alloc.RecentLogs {
-		entry := logs.LogEntry{
-			Instance:  e.Instance,
-			Level:     e.Level,
-			Message:   e.Message,
-			Region:    e.Region,
-			Timestamp: e.Timestamp,
-			Meta:      e.Meta,
-		}
-
-		logPresenter.FPrint(out, cfg, entry)
-	}
 }
