+++ internal/cli/internal/command/image/update.go
@@ -1,0 +1,125 @@
+package image
+
+import (
+	"context"
+	"errors"
+	"fmt"
+
+	"github.com/spf13/cobra"
+
+	"github.com/superfly/flyctl/pkg/iostreams"
+
+	"github.com/superfly/flyctl/api"
+	"github.com/superfly/flyctl/internal/cli/internal/app"
+	"github.com/superfly/flyctl/internal/cli/internal/command"
+	"github.com/superfly/flyctl/internal/cli/internal/flag"
+	"github.com/superfly/flyctl/internal/cli/internal/prompt"
+	"github.com/superfly/flyctl/internal/cli/internal/render"
+	"github.com/superfly/flyctl/internal/cli/internal/watch"
+	"github.com/superfly/flyctl/internal/client"
+)
+
+func newUpdate() *cobra.Command {
+	const (
+		long = `This will update the application's image to the latest available version.
+The update will perform a rolling restart against each VM, which may result in a brief service disruption.
+		`
+		short = "Updates the app's image to the latest available version. (Fly Postgres only)"
+
+		usage = "update"
+	)
+
+	cmd := command.New(usage, short, long, runUpdate,
+		command.RequireSession,
+		command.RequireAppName,
+	)
+
+	flag.Add(
+		cmd,
+		flag.App(),
+		flag.AppConfig(),
+		flag.Bool{
+			Name:        "detach",
+			Description: "Return immediately instead of monitoring update progress",
+		},
+	)
+
+	return cmd
+}
+
+func runUpdate(ctx context.Context) error {
+	var (
+		client  = client.FromContext(ctx).API()
+		io      = iostreams.FromContext(ctx)
+		appName = app.NameFromContext(ctx)
+	)
+
+	app, err := client.GetImageInfo(ctx, appName)
+	if err != nil {
+		return err
+	}
+
+	if !app.ImageVersionTrackingEnabled {
+		return errors.New("image is not eligible for automated image updates")
+	}
+
+	if !app.ImageUpgradeAvailable {
+		return errors.New("image is already running the latest image")
+	}
+
+	cI := app.ImageDetails
+	lI := app.LatestImageDetails
+
+	current := fmt.Sprintf("%s:%s", cI.Repository, cI.Tag)
+	target := fmt.Sprintf("%s:%s", lI.Repository, lI.Tag)
+
+	if cI.Version != "" {
+		current = fmt.Sprintf("%s %s", current, cI.Version)
+	}
+
+	if lI.Version != "" {
+		target = fmt.Sprintf("%s %s", target, lI.Version)
+	}
+
+	if !flag.GetYes(ctx) {
+
+		msg := fmt.Sprintf("Update `%s` from %s to %s?", appName, current, target)
+		if confirmed, err := prompt.Confirm(ctx, msg); err != nil || !confirmed {
+			return err
+		}
+	}
+
+	input := api.DeployImageInput{
+		AppID:    appName,
+		Image:    fmt.Sprintf("%s:%s", lI.Repository, lI.Tag),
+		Strategy: api.StringPointer("ROLLING"),
+	}
+
+	release, releaseCommand, err := client.DeployImage(ctx, input)
+	if err != nil {
+		return err
+	}
+
+	fmt.Fprintf(io.Out, "Release v%d created\n", release.Version)
+	if releaseCommand != nil {
+		fmt.Fprintf(io.Out, "Release command detected: this new release will not be available until the command succeeds.\n")
+	}
+
+	fmt.Fprintln(io.Out)
+
+	tb := render.NewTextBlock(ctx)
+
+	tb.Detail("You can detach the terminal anytime without stopping the update")
+
+	if releaseCommand != nil {
+		// TODO: don't use text block here
+		tb := render.NewTextBlock(ctx, fmt.Sprintf("Release command detected: %s\n", releaseCommand.Command))
+		tb.Done("This release will not be available until the release command succeeds.")
+
+		if err := watch.WatchReleaseCommand(ctx, releaseCommand.ID); err != nil {
+			return err
+		}
+	}
+
+	return watch.WatchDeployment(ctx)
+}
